Move Error and HttpResponses
