+++ identity/app/formstack/FormstackApi.scala
@@ -2,8 +2,9 @@ package formstack
 
 import com.gu.identity.model.LiftJsonConfig
 import conf.Configuration
-import idapiclient.{Error, Response}
+import idapiclient.Response
 import idapiclient.parser.JodaJsonSerializer
+import idapiclient.responses.Error
 import net.liftweb.json._
 import utils.SafeLogging
