+++ identity/app/idapiclient/package.scala
@@ -1,3 +1,5 @@
+import idapiclient.responses.Error
+
 package object idapiclient {
   type Response[T] = Either[List[Error], T]
   type Parameters = Iterable[(String, String)]
