+++ identity/app/idapiclient/responses/Error.scala
@@ -1,3 +1,3 @@
-package idapiclient
+package idapiclient.responses
 
 case class Error(message: String, description: String, statusCode: Int = 500, context: Option[String] = None)
