+++ identity/app/idapiclient/responses/HttpResponse.scala
@@ -1,3 +1,3 @@
-package idapiclient
+package idapiclient.responses
 
 case class HttpResponse(body: String, statusCode: Int, statusMessage: String)
