+++ identity/app/views/password/requestPasswordReset.scala.html
@@ -1,7 +1,7 @@
 @(page: model.Page, idRequest: services.IdentityRequest, idUrlBuilder: services.IdentityUrlBuilder, emailForm: Form[(String)], errors: List[Error])(implicit request: RequestHeader, messages: play.api.i18n.Messages, context: model.ApplicationContext)
 
 @import form.IdFormHelpers._
-@import idapiclient.Error
+@import idapiclient.responses.Error
 @import views.html.fragments.form.inputField
 @import views.html.fragments.registrationFooter
