+++ identity/app/views/profile/deletion/accountDeletion.scala.html
@@ -1,4 +1,5 @@
-@import idapiclient.Error
+@import idapiclient.responses.Error
+
 @(
     page: model.Page,
     idRequest: services.IdentityRequest,
